Separate API to its own file
