+++ server/src/server.js
@@ -1,17 +1,8 @@
 import express from 'express'
-import childProcess from 'child_process'
 import { checkInt, checkBool } from './check'
 import { config } from './setup'
 
-// Run a process `command` and asynchronously return standard outputs.
-function execProcess(command) {
-  return new Promise((resolve, reject) => {
-    childProcess.exec(command, (error, stdout, stderr) => {
-      if (error) reject(error)
-      else resolve({ stdout, stderr })
-    })
-  })
-}
+import apiVersion from './api/version'
 
 export function startServer() {
 
@@ -23,15 +14,7 @@ export function startServer() {
     app.use(express.static('../misc'))
   }
 
-  app.get('/api/version', async (req, res) => {
-    const { stdout: hash } = await execProcess('git rev-parse HEAD')
-    const { stdout: branch } = await execProcess('git rev-parse --abbrev-ref HEAD')
-
-    res.send({
-      gitHash: hash.trim(),
-      gitBranch: branch.trim(),
-    })
-  })
+  app.use(apiVersion)
 
   const port = checkInt('port', config.port)
   app.listen(port, () => {
